+++ src/repositories/provider.rs
@@ -27,10 +27,10 @@ pub trait MetadataProvider {
 /// Generic prebuild repository provider trait, reading prebuild packages from a repository.
 pub trait PrebuildProvider {
     /// Gets the url of a prebuild package, returns None if the prebuild package does not exist
-    fn get_prebuild_url(&self, package_id: &PackageId, revision: usize, target: &str) -> Option<String>;
+    fn get_prebuild_url(&self, package_id: &PackageId, revision: u64, target: &str) -> Option<String>;
 
     /// Gets the checksum of a prebuild package, returns None if the prebuild package does not exist
-    fn get_prebuild_checksum(&self, package_id: &PackageId, revision: usize, target: &str) -> Result<Option<Checksum>>;
+    fn get_prebuild_checksum(&self, package_id: &PackageId, revision: u64, target: &str) -> Result<Option<Checksum>>;
 }
 
 /// Creates a metadata repository provider for the given repository.
